split entry manager into multiple modules
